requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.1
django-crontab==0.7.1
django-webpack-loader==0.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==13.15.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
